+++ dio/lib/src/adapters/browser_adapter.dart
@@ -278,7 +278,9 @@ class BrowserHttpClientAdapter implements HttpClientAdapter {
       }
       final completer = Completer<Uint8List>();
       final sink = ByteConversionSink.withCallback(
-        (bytes) => completer.complete(Uint8List.fromList(bytes)),
+        (bytes) => completer.complete(
+          bytes is Uint8List ? bytes : Uint8List.fromList(bytes),
+        ),
       );
       requestStream.listen(
         sink.add,