+++ plugins/native_dio_adapter/lib/src/conversion_layer_adapter.dart
@@ -56,7 +56,9 @@ class ConversionLayerAdapter implements HttpClientAdapter {
     if (requestStream != null) {
       final completer = Completer<Uint8List>();
       final sink = ByteConversionSink.withCallback(
-        (bytes) => completer.complete(Uint8List.fromList(bytes)),
+        (bytes) => completer.complete(
+          bytes is Uint8List ? bytes : Uint8List.fromList(bytes),
+        ),
       );
       requestStream.listen(
         sink.add,